Add files via upload
Consider a snail travels up a tree a feet each day. Then snail slides down b feet each night. Height of the tree is h feet.
Please, proceed to Snail class
and write a program that prints number of days for the snail to reach the top of the tree.
Program reads a, b, h line by line. Input values are guaranteed to be positive integers.
If the snail cannot reach the top of the tree, print the message Impossible.

Examples

Input:

4
2
14


Output:

6



Input:

4
3
10


Output:

7



Input:

4
4
10


Output:

Impossible



Input:

4
4
1


Output:

1